Auto generate all game events via a script. (#1)
* Add examples of all known journal event types.

+++ src/state/engineering.rs
@@ -10,6 +10,6 @@ pub struct Engineer {
     pub engineer: String,
     pub engineer_id: u64,
     pub progress: String,
-    pub rank_progress: u8,
-    pub rank: u8,
+    pub rank_progress: u64,
+    pub rank: u64,
 }
